requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.17.105
botocore==1.20.105
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.5
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
MarkupSafe==2.0.1
packaging==21.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
